Lots of changes, mostly firmware driver work

+++ Software/Battery Balancer/Project Includes/CAN.h
@@ -56,16 +56,16 @@
 
 typedef struct
 {
-	Uint8 CellSel1;
-	Uint8 CellSel2;
-	Uint8 CellSel3;
-	Uint8 CellSel4;
+	uint8_t CellSel1;
+	uint8_t CellSel2;
+	uint8_t CellSel3;
+	uint8_t CellSel4;
 } cells_t;
 
 typedef struct
 {
-	Uint8 ID1;
-	Uint8 ID2;
+	Uint32 ID1;
+	Uint32 ID2;
 	cells_t Active_Cells;
 	Bool ID1_Active;
 } can_bim_mailbox_t;
